# src/main/java/com/bank/dto/AuthRequest.java
package com.bank.dto;

import com.bank.validation.StrongPassword;
import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class AuthRequest {
    @NotBlank(message = "Username is required")
    private String username;

    @NotBlank(message = "Password is required")
    @StrongPassword
    private String password;
}